* Docs and moved ErrorAmplificationAnalysis
