+++ qiskit_experiments/curve_analysis/standard_analysis/error_amplification_analysis.py
